:recycle: refator: move activity serializer to src/school domain
