+++ api/src/school/application/assessment-controller.js
@@ -1,7 +1,7 @@
 import * as challengeSerializer from '../../shared/infrastructure/serializers/jsonapi/challenge-serializer.js';
 import { usecases } from '../domain/usecases/index.js';
 import * as assessmentSerializer from '../../../src/school/infrastructure/serializers/assessment.js';
-import * as activitySerializer from '../../../lib/infrastructure/serializers/jsonapi/activity-serializer.js';
+import * as activitySerializer from '../infrastructure/serializers/activity-serializer.js';
 
 const getNextChallengeForPix1d = async function (request, h, dependencies = { challengeSerializer }) {
   const assessmentId = request.params.id;
