+++ api/src/school/infrastructure/serializers/activity-serializer.js
